VC7 catching up

+++ ql/MarketModels/driftcalculator.cpp
@@ -61,7 +61,7 @@ void DriftCalculator::compute(const Array& forwards, Array& drifts) const {
 		drifts[i]=0.;
 		Size down = std::min(i+1, numeraire_);
 		Size up = std::max(i+1, numeraire_) - 1;
-		int sign = (i+1>numeraire_ ? +1.0 : -1.0);
+		int sign = (i+1>numeraire_ ? static_cast<int>(+1.0) : static_cast<int>(-1.0));
 		for(Size k=down; k<up; ++k) {
 			const double A = (taus_[k]*(forwards[k]+displacements_[k]) / (1.+taus_[k]*forwards[k]));
 			drifts[i] += sign * A*C_[k][i];
